+++ packages/plugin-auth/src/runtime/index.tsx
@@ -22,7 +22,6 @@ const runtime: RuntimePlugin = async ({ appContext, useConfig, addProvider, addW
   };
 
   const AuthRouteWrapper: RouteWrapper = ({ children }) => {
-    console.log(123123);
     const [auth] = useAuth();
     const routeConfig = useConfig();
     const routeConfigAuth = routeConfig?.auth;