+++ packages/runtime/src/runtime.tsx
@@ -16,7 +16,8 @@ import type {
   ComponentWithChildren,
 } from './types.js';
 import DefaultAppRouter from './AppRouter.js';
-import { useData, useConfig, useRouteModule } from './RouteContext.js';
+import { useData, useConfig } from './RouteContext.js';
+import { useAppContext } from './AppContext.js';
 
 class Runtime {
   private appContext: AppContext;
@@ -59,7 +60,7 @@ class Runtime {
       setAppRouter: this.setAppRouter,
       useData,
       useConfig,
-      useRouteModule,
+      useAppContext,
     };
 
     const runtimeModule = (module as CommonJsRuntime).default || module as RuntimePlugin;