+++ packages/plugin-store/CHANGELOG.md
@@ -1,3 +1,5 @@
 # CHANGELOG
 
 ## 1.0.0
+
+- feat: init plugin store